Migrate to GridSuite dependencies 22 (#212)

+++ src/main/java/org/gridsuite/modification/server/modifications/GeneratorModification.java
@@ -307,7 +307,7 @@ public class GeneratorModification extends AbstractModification {
         List<Report> reports = new ArrayList<>();
 
         ActivePowerControl<Generator> activePowerControl = generator.getExtension(ActivePowerControl.class);
-        Float oldDroop = activePowerControl != null ? activePowerControl.getDroop() : Float.NaN;
+        double oldDroop = activePowerControl != null ? activePowerControl.getDroop() : Double.NaN;
         Boolean participate = null;
         // if participate is null and droop was modified, we consider that participate
         // is true
